requirements.txt
asgiref==3.9.1
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
asgiref==3.9.1
decouple==0.0.7
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
asgiref==3.9.1
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
asgiref==3.9.1
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
newsapi-python==0.2.7
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-debug-toolbar==6.0.0
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
newsapi-python==0.2.7
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-debug-toolbar==6.0.0
django-dirtyfields==1.9.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
newsapi-python==0.2.7
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
